feat: rename trainer config package to config
